[CLI] Butler commands TBI (#5032)
This PR adds the `butler.py` commands to the `casp` CLI, but leaves it
marked as "To be implemented".

These changes are intended to serve as a roadmap for further
implementation. Since the commands in `butler.py` need to be validated
before migrating to the new CLI, it's crucial to establish a structure
and expectations around which commands should be included in `casp`.
This will serve as a reminder that these commands should be incorporated
into the CLI.

Those commands are:
- [ ] bootstrap           
- [ ] py_unittest         
- [ ] js_unittest         
- [ ] format              
- [ ] lint                
- [ ] package             
- [ ] deploy              
- [ ] run_server          
- [ ] run                 
- [ ] run_bot             
- [ ] remote              
- [ ] clean_indexes       
- [ ] create_config       
- [ ] integration_tests   
- [ ] weights             
- [x] reproduce