Add log to analyze FuzzTargetJob entity deletion (#5051)
The cleanup cronjob is failing for Internal due to a `FuzzTargetJob`
query timeout, and the discussion in b/418807403 indicates that the
issue may be related to the deletion of `FuzzTargetJob` entities. This
PR adds a log to help confirm if deletion is the main problem, which
will be useful for developing an appropriate solution.

Key changes:
- Added a log to indicate how many `FuzzTargetJob` entities are deleted
per cronjob and how many remain in the database.
- Transformed the query iterator into a list to log the remaining
entities. This may increase memory usage, but it should not be a problem
as this kind has around 5,000 entities.

Related to b/418807403 and b/406250205.