[CLI] remove js_unittest from casp (#5058)
This PR removes the `js_unittest` command and its associated
implementation from `casp`.

### Reasoning
The `js_unittest` command is deprecated and relies on significantly
outdated versions of Chrome and Chromedriver. As noted in [this
issue](https://github.com/google/clusterfuzz/issues/2678), the current
implementation is unstable/broken. [This
PR](https://github.com/google/clusterfuzz/pull/2689) commented out the
call to this command a while ago.

Fixing this functionality would require a complex migration to the new
Chrome/Chromedriver versioning system, which is described
[here](https://developer.chrome.com/docs/chromedriver/downloads/version-selection).
Since the command is already deprecated, we have decided to remove it
rather than investing resources in the upgrade for now.

### Changes
Removed the `js_unittest` source file.

Removed the command registration and module calls in the main script.