Assign predator bug component id even when missing suspected components (#5065)
Recently, a possible issue in predator was making testcases miss the
`suspected_components` field in predator result, even though the correct
`suspected_buganizer_component_id` was set. However, the logic in
cleanup was preventing the bugs to be assigned with this component ID.
This PR fixes by allowing this case.

Context: b/464899796

---------

Co-authored-by: Titouan Rigoudy <titouan@chromium.org>