[batch] Add congestion jobs.
The idea is if we schedule a bunch of NOOP jobs and we don't see
many in the last hour, that things are congested.
This might not react fast enough to congestion, but can potentially
pause things for long enough to drain the queue.
In the future we should stop hardcoding and infer how many we should
see completed in the last hour.
Other future improvements can be testing congestion by subconfig
so we can tell if some other resource is saturated.