Create remote task interface (#5082)
Fix  b/469102717

It creates the Remote Task interface for running tasks remotely as a job
on different infrastructures such as GCP Batch, Kubernetes, Cloud Run
and so on.

For now this PR creates this abstraction with the `create_job` function
as the shared contract for the interface.

It also move the GCP Batch implementation to behind of this interface.